backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.4
python-dotenv==1.0.0
Werkzeug==2.3.7
redis==5.0.0
celery==5.3.1